+++ lib/src/snippet/snodes/chip_node.dart
@@ -284,6 +284,7 @@ class ChipNode extends CL with ChipNodeMappable {
       fco.addSubRoute(
           newPath: destinationRoutePathSnippetName!,
           template: SnippetTemplateEnum.empty);
+      fco.pageList.add(destinationRoutePathSnippetName!);
       context.replace(destinationRoutePathSnippetName!);
       // create a GoRoute and load or create snippet with pageName
     } else if (destinationPanelOrPlaceholderName != null &&