+++ lib/x_flutter_content/routes_extn.dart
@@ -20,7 +20,6 @@ extension RoutesExtension on FlutterContentMixins {
     List<RouteBase> subRoutes = routingConfigVN.value.routes;
     if (!newPath.endsWith(' missing')) {
       subRoutes.add(DynamicPageRoute(path: newPath, template: template));
-      pageList.add(newPath);;
     }
   }
 
@@ -54,7 +53,7 @@ extension RoutesExtension on FlutterContentMixins {
           return path.startsWith('/') ? path : '/$path';
         })
         // .toList()
-        // .where((routePath) => !appInfo.sandboxPageNames.contains(routePath))
+        // .where((routePath) => !appInfo.userEditablePages.contains(routePath))
         .toList();
   }
 
@@ -63,10 +62,8 @@ extension RoutesExtension on FlutterContentMixins {
     // RouteBase home = routingConfig.routes.first;
     for (String snippetName in appInfo.snippetNames) {
       if (snippetName.startsWith('/') && !pageList.contains(snippetName)) {
-        addSubRoute(
-          newPath: snippetName,
-          template: SnippetTemplateEnum.empty,
-        );
+        addSubRoute(newPath: snippetName, template: SnippetTemplateEnum.empty);
+        pageList.add(snippetName);
       }
     }
   }
@@ -88,6 +85,7 @@ extension RoutesExtension on FlutterContentMixins {
         String matchedLocation = state.matchedLocation;
         // var param = state.pathParameters;
 
+        // implicit built-in page
         if (matchedLocation == '/pages') {
           if (fco.canEditContent()) {
             return Pages();
@@ -109,103 +107,74 @@ extension RoutesExtension on FlutterContentMixins {
         final snippetNames = appInfo.snippetNames;
         bool dynamicPageExists = snippetNames.contains(matchedLocation);
 
-        // bool contains(list, key) {
-        //   for (String s in list) {
-        //     print('$s, $key');
-        //     if (s == key)
-        //       return true;
-        //   }
-        //   return false;
+        // if (dynamicPageExists) {
+        //   EditablePage.removeAllNodeWidgetOverlays();
+        //   // fco.dismiss('exit-editMode');
+        //   final snippetName = matchedLocation;
+        //   final rootNode = SnippetTemplateEnum.empty.clone()
+        //     ..name = snippetName;
+        //   SnippetRootNode.loadSnippetFromCacheOrFromFBOrCreateFromTemplate(
+        //     snippetName: snippetName,
+        //     templateSnippetRootNode: rootNode,
+        //   );
+        //   final dynamicPage = EditablePage(
+        //     key: GlobalKey(), // provides access to state later
+        //     routePath: matchedLocation,
+        //     child: SnippetPanel.fromSnippet(
+        //       panelName: "dynamic panel",
+        //       snippetName: snippetName,
+        //       scName: null,
+        //       key: ValueKey<String>(snippetName),
+        //     ),
+        //   );
+        //   return dynamicPage;
         // }
-        // dynamicPageExists = contains(snippetNames,matchedLocation);
-        // may still exists if the matchedLocation is in the form: / + callout content id
-        if (!dynamicPageExists) {
-          // final possiblyACalloutContentId = matchedLocation.substring(1);
-          // final isCID = int.tryParse(possiblyACalloutContentId) != null
-          //     || possiblyACalloutContentId.startsWith('T-');
-          // if (isCID) {
-          //   final cid = possiblyACalloutContentId;
-          //   if (snippetNames.contains(cid)) {
-          //     return CalloutContentEditablePage(
-          //       key: GlobalKey(), // provides access to state later
-          //       cid: cid,
-          //       child: SnippetPanel.fromSnippet(
-          //         panelName: "callout-content-editor-panel",
-          //         snippetName: cid,
-          //         scName: cid,
-          //       ),
-          //     );
-          //   }
-          // }
-        }
-        if (dynamicPageExists) {
-          EditablePage.removeAllNodeWidgetOverlays();
-          // fco.dismiss('exit-editMode');
-          final snippetName = matchedLocation;
-          final rootNode = SnippetTemplateEnum.empty.clone()
-            ..name = snippetName;
-          SnippetRootNode.loadSnippetFromCacheOrFromFBOrCreateFromTemplate(
-            snippetName: snippetName,
-            templateSnippetRootNode: rootNode,
-          );
-          final dynamicPage = EditablePage(
-            key: GlobalKey(), // provides access to state later
-            routePath: matchedLocation,
-            child: SnippetPanel.fromSnippet(
-              panelName: "dynamic panel",
-              snippetName: snippetName,
-              scName: null,
-              key: ValueKey<String>(snippetName),
-            ),
-          );
-          return dynamicPage;
-        }
 
-        // page doesn't exist yet
-        // editor can ask to create it
-        if (canEditContent()) {
-          return AlertDialog(
-            title: Text('Page "$matchedLocation" does not Exist !'),
-            content: SingleChildScrollView(
-              child: ListBody(
-                children: const <Widget>[Text('Want to create it now ?')],
-              ),
-            ),
-            actions: <Widget>[
-              TextButton(
-                child: Text('Yes, Create page $matchedLocation'),
-                onPressed: () {
-                  final String destUrl = matchedLocation;
-                  EditablePage.removeAllNodeWidgetOverlays();
-                  // fco.dismiss('exit-editMode');
-                  // bool userCanEdit = canEditContent.isTrue;
-                  final snippetName = destUrl;
-                  final rootNode = SnippetTemplateEnum.empty.clone()
-                    ..name = snippetName;
-                  SnippetRootNode.loadSnippetFromCacheOrFromFBOrCreateFromTemplate(
-                    snippetName: snippetName,
-                    templateSnippetRootNode: rootNode,
-                  ).then((_) {
-                    afterNextBuildDo(() {
-                      // SnippetInfoModel.snippetInfoCache;
-                      router.push(destUrl);
-                      // router.go('/');
-                    });
-                  });
-                },
-              ),
-              TextButton(
-                child: const Text('Cancel'),
-                onPressed: () {
-                  context.go('/');
-                },
-              ),
-            ],
-          );
-        }
+        // // page doesn't exist yet
+        // // editor can ask to create it
+        // if (canEditContent()) {
+        //   return AlertDialog(
+        //     title: Text('Page "$matchedLocation" does not Exist !'),
+        //     content: SingleChildScrollView(
+        //       child: ListBody(
+        //         children: const <Widget>[Text('Want to create it now ?')],
+        //       ),
+        //     ),
+        //     actions: <Widget>[
+        //       TextButton(
+        //         child: Text('Yes, Create page $matchedLocation'),
+        //         onPressed: () {
+        //           final String destUrl = matchedLocation;
+        //           EditablePage.removeAllNodeWidgetOverlays();
+        //           // fco.dismiss('exit-editMode');
+        //           // bool userCanEdit = canEditContent.isTrue;
+        //           final snippetName = destUrl;
+        //           final rootNode = SnippetTemplateEnum.empty.clone()
+        //             ..name = snippetName;
+        //           SnippetRootNode.loadSnippetFromCacheOrFromFBOrCreateFromTemplate(
+        //             snippetName: snippetName,
+        //             templateSnippetRootNode: rootNode,
+        //           ).then((_) {
+        //             afterNextBuildDo(() {
+        //               // SnippetInfoModel.snippetInfoCache;
+        //               router.push(destUrl);
+        //               // router.go('/');
+        //             });
+        //           });
+        //         },
+        //       ),
+        //       TextButton(
+        //         child: const Text('Cancel'),
+        //         onPressed: () {
+        //           context.go('/');
+        //         },
+        //       ),
+        //     ],
+        //   );
+        // }
 
-        // a sandbox page has been created
-        if (fco.appInfo.sandboxPageNames.contains(matchedLocation)) {
+        // dynamic user-editable page
+        /*    if (fco.appInfo.userEditablePages.contains(matchedLocation)) {
           final String destUrl = matchedLocation;
           EditablePage.removeAllNodeWidgetOverlays();
           // fco.dismiss('exit-editMode');
@@ -224,42 +193,18 @@ extension RoutesExtension on FlutterContentMixins {
             });
           });
         }
-
+*/
         return AlertDialog(
           title: const Text('Page does not Exist !'),
-          content: SingleChildScrollView(
-            child: ListBody(
-              children: <Widget>[
-                Column(
-                  mainAxisSize: MainAxisSize.max,
-                  children: [
-                    Text(
-                      'You must be signed in as an editor to create this page.',
-                    ),
-                    Gap(10),
-                    ElevatedButton(
-                      onPressed: () {
-                        context.go('/');
-                      },
-                      child: const Text('ok'),
-                    ),
-                  ],
-                ),
-              ],
-            ),
+          content: ElevatedButton(
+            onPressed: () {
+              context.go('/');
+            },
+            child: const Text('go back'),
           ),
-          // actions: <Widget>[
-          //   TextButton(
-          //     child: const Text('ok'),
-          //     onPressed: () {
-          //       context.go('/');
-          //     },
-          //   ),
-          // ],
         );
       },
       observers: [GoRouterObserver()],
     );
-
   }
 }